Added and updated weapons files
Added weapons files for the various Hellfire II iterations - this covers fallbacks for all 3 Apache variants, the Supercobra, and the Kiowa.  This also adds the 184 Long and 131 pods. Lastly, this adds date and fallback information to the shrikes in advance of the AGM-45B being added to the A-4E mod.