F-16A File Updates
Updates the filename of the F-16A banner so that liberation can read it.

Updates the F-16A.yaml resource file to add BAI / CAS / antiship mission types - since the 15A is capable of these.

Updated the F-16A payload provide more capability - primarily adding jamming pods to the centerline, moving the fuel tanks from the centerline to the inner wings, replacing the AIS_ASQ_T50 on the wingtips with 120Bs, and changing the CAP loadout to have 120Bs on pylons 8 and 3 so they fall back to sparrows on historical campaigns.