Add generation-time allocation for STNs.
This is a stop-gap solution so we can get a release out that supports
datalink as quickly as possible. It would be better to assign these
early so the player can select which flights should be their
team/donors, but for that we need to keep a registry in Game, plumb it
through to each Flight creation (and there are many callsites for that),
subscribe each flight to its datalink contacts (so disbanded flights do
not remain part of a network), and of course the UI itself. That'll come
later.

Follow up work:

* Filtering to only apply for the airframes that can use it
* Selection of team/donors